[refac] Move metric wrapper to separate module
