+++ metrics/Cargo.toml
@@ -17,10 +17,6 @@ include = [ "Cargo.toml", "src", "README.md", "LICENSE.md" ]
 license = "GPL-3.0"
 edition = "2018"
 
-[dependencies.const-default]
-version = "0.1"
-features = ["derive"]
-
 [dependencies.futures-util]
 version = "0.3.5"
 default-features = false
